Mini-Game-Management-Platform/requirements.txt
fastapi
uvicorn
mysql-connector-python
passlib
bcrypt==4.0.1
